DOC/MNT : moved RCParamsDict to core
