+++ doc/api_changes/2014-09-30-rcparams.rst
@@ -1,0 +1,3 @@
+Changes
+-------
+Moved RCPParamsDict to core and removed rcparams.py
