--- doc/resource/api/rcparams.rst
@@ -1,9 +1,0 @@
-========================
- :mod:`rcparams` Module
-========================
-
-
-.. automodule:: nsls2.rcparams
-   :members:
-   :show-inheritance:
-   :undoc-members:
